3.24.2 Advanced VKB Gladiator NXT Binds
Uploading my 3.24.2 Advanced NXT binds to try out this open source idea. This could be the start of something magical. 
Side note, there is a dialogue box above this one that says "add files via upload" and I have no idea what I am supposed to type there... it isn't a thing you click on to upload the files it is a box that I would have suspected was supposed to the title to whatever this message I am typing is.